test(IDX-3144): finish inlining of system-tests
