+++ rs/tests/BUILD.bazel
@@ -4,9 +4,7 @@ load("@bazel_skylib//rules:run_binary.bzl", "run_binary")
 load("@rules_distroless//apt:defs.bzl", "dpkg_status")
 load("@rules_distroless//distroless:defs.bzl", "passwd")
 load("@rules_oci//oci:defs.bzl", "oci_image")
-load("@rules_rust//rust:defs.bzl", "rust_library")
 load("//bazel:defs.bzl", "symlink_dir", "symlink_dir_test")
-load("//rs/tests:common.bzl", "DEPENDENCIES", "MACRO_DEPENDENCIES")
 load(":system_tests.bzl", "oci_tar", "uvm_config_image")
 
 package(default_visibility = ["//rs:system-tests-pkg"])
@@ -18,19 +16,6 @@ exports_files([
 bool_flag(
     name = "k8s",
     build_setting_default = False,
-)
-
-rust_library(
-    name = "tests",
-    testonly = True,
-    srcs = glob(["src/**/*.rs"]),
-    compile_data = [
-        "//ic-os/components:ic/generate-ic-config/ic.json5.template",
-    ],
-    crate_name = "ic_tests",
-    proc_macro_deps = MACRO_DEPENDENCIES,
-    target_compatible_with = ["@platforms//os:linux"],  # requires libssh that does not build on Mac OS
-    deps = DEPENDENCIES,
 )
 
 # Packages we install into the image; see base image definition and its
