+++ rs/tests/consensus/subnet_recovery/BUILD.bazel
@@ -1,5 +1,5 @@
 load("@rules_rust//rust:defs.bzl", "rust_library")
-load("//rs/tests:common.bzl", "GUESTOS_RUNTIME_DEPS", "MACRO_DEPENDENCIES", "UNIVERSAL_VM_RUNTIME_DEPS")
+load("//rs/tests:common.bzl", "GUESTOS_RUNTIME_DEPS", "UNIVERSAL_VM_RUNTIME_DEPS")
 load("//rs/tests:system_tests.bzl", "system_test_nns")
 
 package(default_visibility = ["//rs:system-tests-pkg"])
@@ -49,7 +49,6 @@ system_test_nns(
     name = "sr_app_same_nodes_test",
     extra_head_nns_tags = [],  # don't run the head_nns variant on nightly since it aleady runs on long_test.
     flaky = True,
-    proc_macro_deps = MACRO_DEPENDENCIES,
     tags = [
         "experimental_system_test_colocation",
         "k8s",
@@ -69,7 +68,6 @@ system_test_nns(
 system_test_nns(
     name = "sr_app_same_nodes_enable_tecdsa_test",
     flaky = True,
-    proc_macro_deps = MACRO_DEPENDENCIES,
     tags = [
         "experimental_system_test_colocation",
         "k8s",
@@ -89,7 +87,6 @@ system_test_nns(
 system_test_nns(
     name = "sr_app_same_nodes_with_tecdsa_test",
     flaky = True,
-    proc_macro_deps = MACRO_DEPENDENCIES,
     tags = [
         "experimental_system_test_colocation",
         "k8s",
@@ -110,7 +107,6 @@ system_test_nns(
     name = "sr_app_failover_nodes_test",
     extra_head_nns_tags = [],  # don't run the head_nns variant on nightly since it aleady runs on long_test.
     flaky = True,
-    proc_macro_deps = MACRO_DEPENDENCIES,
     tags = [
         "experimental_system_test_colocation",
         "k8s",
@@ -130,7 +126,6 @@ system_test_nns(
 system_test_nns(
     name = "sr_app_failover_nodes_enable_tecdsa_test",
     flaky = True,
-    proc_macro_deps = MACRO_DEPENDENCIES,
     tags = [
         "experimental_system_test_colocation",
         "k8s",
@@ -151,7 +146,6 @@ system_test_nns(
 system_test_nns(
     name = "sr_app_failover_nodes_with_tecdsa_test",
     flaky = True,
-    proc_macro_deps = MACRO_DEPENDENCIES,
     tags = [
         "experimental_system_test_colocation",
         "k8s",
@@ -173,7 +167,6 @@ system_test_nns(
     name = "sr_app_no_upgrade_test",
     extra_head_nns_tags = [],  # don't run the head_nns variant on nightly since it aleady runs on long_test.
     flaky = True,
-    proc_macro_deps = MACRO_DEPENDENCIES,
     tags = [
         "experimental_system_test_colocation",
         "k8s",
@@ -193,7 +186,6 @@ system_test_nns(
 system_test_nns(
     name = "sr_app_no_upgrade_enable_tecdsa_test",
     flaky = True,
-    proc_macro_deps = MACRO_DEPENDENCIES,
     tags = [
         "experimental_system_test_colocation",
         "k8s",
@@ -214,7 +206,6 @@ system_test_nns(
 system_test_nns(
     name = "sr_app_no_upgrade_with_tecdsa_test",
     flaky = True,
-    proc_macro_deps = MACRO_DEPENDENCIES,
     tags = [
         "experimental_system_test_colocation",
         "k8s",
@@ -236,7 +227,6 @@ system_test_nns(
     name = "sr_app_large_with_tecdsa_test",
     extra_head_nns_tags = ["manual"],  # Let's not run this expensive test against the HEAD NNS canisters to save resources.
     flaky = True,
-    proc_macro_deps = MACRO_DEPENDENCIES,
     tags = [
         "experimental_system_test_colocation",
         "subnet_recovery",
@@ -258,7 +248,6 @@ system_test_nns(
     name = "sr_nns_same_nodes_test",
     extra_head_nns_tags = [],  # don't run the head_nns variant on nightly since it aleady runs on long_test.
     flaky = True,
-    proc_macro_deps = MACRO_DEPENDENCIES,
     tags = [
         "experimental_system_test_colocation",
         "k8s",
@@ -284,7 +273,6 @@ system_test_nns(
     name = "sr_nns_failover_nodes_test",
     extra_head_nns_tags = [],  # don't run the head_nns variant on nightly since it aleady runs on long_test.
     flaky = False,
-    proc_macro_deps = MACRO_DEPENDENCIES,
     tags = [
         "experimental_system_test_colocation",
         "k8s",
