+++ rs/tests/consensus/upgrade/BUILD.bazel
@@ -1,5 +1,5 @@
 load("@rules_rust//rust:defs.bzl", "rust_library")
-load("//rs/tests:common.bzl", "GUESTOS_RUNTIME_DEPS", "MACRO_DEPENDENCIES", "MAINNET_REVISION_RUNTIME_DEPS", "UNIVERSAL_VM_RUNTIME_DEPS")
+load("//rs/tests:common.bzl", "GUESTOS_RUNTIME_DEPS", "MAINNET_REVISION_RUNTIME_DEPS", "UNIVERSAL_VM_RUNTIME_DEPS")
 load("//rs/tests:system_tests.bzl", "system_test_nns")
 
 package(default_visibility = ["//rs:system-tests-pkg"])
@@ -29,7 +29,6 @@ system_test_nns(
     colocated_test_driver_vm_resources = {
         "vcpus": 24,
     },
-    proc_macro_deps = MACRO_DEPENDENCIES,
     tags = [
         "experimental_system_test_colocation",
         "system_test_hourly",
@@ -60,7 +59,6 @@ system_test_nns(
 system_test_nns(
     name = "upgrade_downgrade_nns_subnet_test",
     flaky = True,
-    proc_macro_deps = MACRO_DEPENDENCIES,
     tags = [
         "system_test_hourly",
     ],
@@ -86,7 +84,6 @@ system_test_nns(
     name = "upgrade_app_subnet_test",
     extra_head_nns_tags = [],  # don't run the head_nns variant on nightly since it aleady runs on long_test.
     flaky = True,
-    proc_macro_deps = MACRO_DEPENDENCIES,
     tags = [
         "long_test",  # since it takes longer than 5 minutes.
     ],
@@ -113,7 +110,6 @@ system_test_nns(
     name = "downgrade_app_subnet_test",
     extra_head_nns_tags = [],  # don't run the head_nns variant on nightly since it aleady runs on long_test.
     flaky = True,
-    proc_macro_deps = MACRO_DEPENDENCIES,
     tags = [
         "long_test",  # since it takes longer than 5 minutes.
     ],
@@ -139,7 +135,6 @@ system_test_nns(
 system_test_nns(
     name = "upgrade_with_alternative_urls",
     flaky = True,
-    proc_macro_deps = MACRO_DEPENDENCIES,
     tags = [
         "k8s",
         "system_test_hourly",
@@ -163,7 +158,6 @@ system_test_nns(
 system_test_nns(
     name = "unassigned_node_upgrade_test",
     flaky = True,
-    proc_macro_deps = MACRO_DEPENDENCIES,
     tags = [
         "k8s",
         "system_test_hourly",
