+++ rs/tests/financial_integrations/rosetta/rosetta_test_lib/mod.rs
@@ -1,7 +1,6 @@
 mod governance_client;
 pub mod ledger_client;
-pub mod lib;
 pub mod rosetta_client;
 pub mod setup;
 pub mod test_neurons;
-pub mod tests;
+pub mod utils;
