+++ rs/tests/financial_integrations/rosetta/rosetta_test_lib/rosetta_client.rs
@@ -1,10 +1,8 @@
 use ic_base_types::CanisterId;
 use ic_ledger_core::block::BlockIndex;
 use ic_nns_governance_api::pb::v1::Proposal;
-use ic_rosetta_api::convert::to_model_account_identifier;
-use ic_system_test_driver::driver::resource::AllocatedVm;
-
 use ic_rosetta_api::{
+    convert::to_model_account_identifier,
     ledger_client::pending_proposals_response::PendingProposalsResponse,
     models::{
         AccountBalanceMetadata, AccountBalanceRequest, AccountBalanceResponse, AccountType,
@@ -21,6 +19,7 @@ use ic_rosetta_api::{
     },
     request_types::GetProposalInfo,
 };
+use ic_system_test_driver::driver::resource::AllocatedVm;
 use icp_ledger::AccountIdentifier;
 use rand::{seq::SliceRandom, thread_rng};
 use reqwest::{Client as HttpClient, StatusCode as HttpStatusCode};
