+++ rs/tests/financial_integrations/rosetta/rosetta_test_lib/test_neurons.rs
@@ -1,23 +1,19 @@
-use crate::rosetta_tests::lib::{create_neuron, NeuronDetails};
+use crate::utils::create_custom_neuron;
+use crate::utils::{create_neuron, NeuronDetails};
 use ic_ledger_core::Tokens;
 use ic_nns_governance_api::pb::v1::Neuron;
 use ic_rosetta_test_utils::EdKeypair;
 use icp_ledger::AccountIdentifier;
 use std::collections::{BTreeMap, HashMap};
 
-use super::lib::create_custom_neuron;
-
-pub(crate) struct TestNeurons<'a> {
+pub struct TestNeurons<'a> {
     neurons: BTreeMap<u64, Neuron>,
     seed: u64,
     ledger_balances: &'a mut HashMap<AccountIdentifier, Tokens>,
 }
 
 impl TestNeurons<'_> {
-    pub(crate) fn new(
-        seed: u64,
-        ledger_balances: &mut HashMap<AccountIdentifier, Tokens>,
-    ) -> TestNeurons {
+    pub fn new(seed: u64, ledger_balances: &mut HashMap<AccountIdentifier, Tokens>) -> TestNeurons {
         TestNeurons {
             neurons: BTreeMap::default(),
             seed: seed * 100_000,
@@ -26,14 +22,14 @@ impl TestNeurons<'_> {
     }
 
     /// Add a new test neuron and return the details about created neuron.
-    pub(crate) fn create(&mut self, neuron_setup: impl FnOnce(&mut Neuron)) -> NeuronDetails {
+    pub fn create(&mut self, neuron_setup: impl FnOnce(&mut Neuron)) -> NeuronDetails {
         let details = create_neuron(self.seed, neuron_setup, self.ledger_balances);
         self.neurons
             .insert(details.neuron.id.unwrap().id, details.neuron.clone());
         self.seed += 1;
         details
     }
-    pub(crate) fn create_custom(
+    pub fn create_custom(
         &mut self,
         neuron_setup: impl FnOnce(&mut Neuron),
         id: u64,
@@ -47,7 +43,7 @@ impl TestNeurons<'_> {
     }
 
     /// Return the map of neurons indexed by their id.
-    pub(crate) fn get_neurons(&self) -> BTreeMap<u64, Neuron> {
+    pub fn get_neurons(&self) -> BTreeMap<u64, Neuron> {
         self.neurons.clone()
     }
 }
