+++ rs/tests/nested/BUILD.bazel
@@ -41,7 +41,6 @@ rust_library(
 system_test_nns(
     name = "registration",
     flaky = True,
-    proc_macro_deps = MACRO_DEPENDENCIES,
     tags = [
         "system_test_hourly",
     ],
@@ -56,7 +55,6 @@ system_test_nns(
 system_test_nns(
     name = "upgrade",
     flaky = True,
-    proc_macro_deps = MACRO_DEPENDENCIES,
     tags = [
         "system_test_hourly",
     ],
