+++ rs/tests/query_stats/BUILD.bazel
@@ -1,4 +1,4 @@
-load("//rs/tests:common.bzl", "DEPENDENCIES", "GRAFANA_RUNTIME_DEPS", "GUESTOS_RUNTIME_DEPS", "MACRO_DEPENDENCIES", "UNIVERSAL_CANISTER_ENV", "UNIVERSAL_CANISTER_RUNTIME_DEPS")
+load("//rs/tests:common.bzl", "GRAFANA_RUNTIME_DEPS", "GUESTOS_RUNTIME_DEPS", "UNIVERSAL_CANISTER_ENV", "UNIVERSAL_CANISTER_RUNTIME_DEPS")
 load("//rs/tests:system_tests.bzl", "system_test")
 
 package(default_visibility = ["//rs:system-tests-pkg"])
@@ -7,14 +7,15 @@ system_test(
     name = "query_stats_basic",
     env = UNIVERSAL_CANISTER_ENV,
     flaky = True,
-    proc_macro_deps = MACRO_DEPENDENCIES,
     tags = [
         "k8s",
     ],
     target_compatible_with = ["@platforms//os:linux"],  # requires libssh that does not build on Mac OS
     runtime_deps = GUESTOS_RUNTIME_DEPS + GRAFANA_RUNTIME_DEPS + UNIVERSAL_CANISTER_RUNTIME_DEPS,
-    deps = DEPENDENCIES + [
+    deps = [
+        "//rs/tests/driver:ic-system-test-driver",
         "//rs/tests/query_stats/lib",
+        "@crate_index//:anyhow",
     ],
 )
 
@@ -22,15 +23,16 @@ system_test(
     name = "query_stats_below_threshold",
     env = UNIVERSAL_CANISTER_ENV,
     flaky = True,
-    proc_macro_deps = MACRO_DEPENDENCIES,
     tags = [
         "k8s",
         "long_test",  # since it takes longer than 5 minutes.
     ],
     target_compatible_with = ["@platforms//os:linux"],  # requires libssh that does not build on Mac OS
     runtime_deps = GUESTOS_RUNTIME_DEPS + GRAFANA_RUNTIME_DEPS + UNIVERSAL_CANISTER_RUNTIME_DEPS,
-    deps = DEPENDENCIES + [
+    deps = [
+        "//rs/tests/driver:ic-system-test-driver",
         "//rs/tests/query_stats/lib",
+        "@crate_index//:anyhow",
     ],
 )
 
@@ -38,13 +40,14 @@ system_test(
     name = "query_stats_above_threshold",
     env = UNIVERSAL_CANISTER_ENV,
     flaky = True,
-    proc_macro_deps = MACRO_DEPENDENCIES,
     tags = [
         "k8s",
     ],
     target_compatible_with = ["@platforms//os:linux"],  # requires libssh that does not build on Mac OS
     runtime_deps = GUESTOS_RUNTIME_DEPS + GRAFANA_RUNTIME_DEPS + UNIVERSAL_CANISTER_RUNTIME_DEPS,
-    deps = DEPENDENCIES + [
+    deps = [
+        "//rs/tests/driver:ic-system-test-driver",
         "//rs/tests/query_stats/lib",
+        "@crate_index//:anyhow",
     ],
 )
