--- rs/tests/src/rosetta_tests/tests.rs
@@ -1,2 +1,0 @@
-pub mod neuron_follow;
-pub mod neuron_voting;
