+++ rs/tests/testing_verification/testnets/BUILD.bazel
@@ -1,4 +1,4 @@
-load("//rs/tests:common.bzl", "BOUNDARY_NODE_GUESTOS_RUNTIME_DEPS", "GRAFANA_RUNTIME_DEPS", "GUESTOS_RUNTIME_DEPS", "IC_MAINNET_NNS_RECOVERY_RUNTIME_DEPS", "MACRO_DEPENDENCIES", "NNS_CANISTER_ENV", "NNS_CANISTER_RUNTIME_DEPS", "SNS_CANISTER_ENV", "SNS_CANISTER_RUNTIME_DEPS")
+load("//rs/tests:common.bzl", "BOUNDARY_NODE_GUESTOS_RUNTIME_DEPS", "GRAFANA_RUNTIME_DEPS", "GUESTOS_RUNTIME_DEPS", "IC_MAINNET_NNS_RECOVERY_RUNTIME_DEPS", "NNS_CANISTER_ENV", "NNS_CANISTER_RUNTIME_DEPS", "SNS_CANISTER_ENV", "SNS_CANISTER_RUNTIME_DEPS")
 load("//rs/tests:system_tests.bzl", "system_test", "system_test_nns")
 
 package(default_visibility = ["//rs:system-tests-pkg"])
@@ -8,7 +8,6 @@ package(default_visibility = ["//rs:system-tests-pkg"])
 system_test(
     name = "single_large_node",
     flaky = False,
-    proc_macro_deps = MACRO_DEPENDENCIES,
     tags = [
         "dynamic_testnet",
         "k8s",
@@ -27,7 +26,6 @@ system_test(
 system_test(
     name = "single_app_large_node",
     flaky = False,
-    proc_macro_deps = MACRO_DEPENDENCIES,
     tags = [
         "dynamic_testnet",
         "k8s",
@@ -46,7 +44,6 @@ system_test(
 system_test_nns(
     name = "single_app_large_node_with_nns",
     flaky = False,
-    proc_macro_deps = MACRO_DEPENDENCIES,
     tags = [
         "dynamic_testnet",
         "k8s",
@@ -66,7 +63,6 @@ system_test_nns(
 system_test_nns(
     name = "small",
     flaky = False,
-    proc_macro_deps = MACRO_DEPENDENCIES,
     tags = [
         "dynamic_testnet",
         "manual",
@@ -85,7 +81,6 @@ system_test_nns(
 system_test_nns(
     name = "small_high_perf",
     flaky = False,
-    proc_macro_deps = MACRO_DEPENDENCIES,
     tags = [
         "dynamic_testnet",
         "manual",
@@ -106,7 +101,6 @@ system_test(
     name = "from_config",
     env = NNS_CANISTER_ENV,
     flaky = False,
-    proc_macro_deps = MACRO_DEPENDENCIES,
     tags = [
         "dynamic_testnet",
         "manual",
@@ -124,7 +118,6 @@ system_test(
 system_test_nns(
     name = "small_nns",
     flaky = False,
-    proc_macro_deps = MACRO_DEPENDENCIES,
     tags = [
         "dynamic_testnet",
         "manual",
@@ -149,7 +142,6 @@ system_test_nns(
 system_test_nns(
     name = "small_with_api_bn",
     flaky = False,
-    proc_macro_deps = MACRO_DEPENDENCIES,
     tags = [
         "dynamic_testnet",
         "manual",
@@ -180,7 +172,6 @@ system_test_nns(
         "SUBNET_RENTAL_WASM_PATH": "$(rootpath @subnet_rental_canister//file)",
     },
     flaky = False,
-    proc_macro_deps = MACRO_DEPENDENCIES,
     tags = [
         "dynamic_testnet",
         "manual",
@@ -210,7 +201,6 @@ system_test_nns(
         "SUBNET_RENTAL_WASM_PATH": "$(rootpath @subnet_rental_canister//file)",
     },
     flaky = False,
-    proc_macro_deps = MACRO_DEPENDENCIES,
     tags = [
         "dynamic_testnet",
         "manual",
@@ -236,7 +226,6 @@ system_test_nns(
 system_test_nns(
     name = "medium",
     flaky = False,
-    proc_macro_deps = MACRO_DEPENDENCIES,
     tags = [
         "dynamic_testnet",
         "manual",
@@ -261,7 +250,6 @@ system_test_nns(
         "SUBNET_RENTAL_WASM_PATH": "$(rootpath @subnet_rental_canister//file)",
     },
     flaky = False,
-    proc_macro_deps = MACRO_DEPENDENCIES,
     tags = [
         "dynamic_testnet",
         "manual",
@@ -293,7 +281,6 @@ system_test_nns(
         "SUBNET_RENTAL_WASM_PATH": "$(rootpath @subnet_rental_canister//file)",
     },
     flaky = False,
-    proc_macro_deps = MACRO_DEPENDENCIES,
     tags = [
         "dynamic_testnet",
         "manual",
